requirements.txt
# Core Frameworks & Web UI
streamlit==1.38.0
pandas==2.2.2

# Data Ingestion & Streaming
kafka-python==2.0.3
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3

# RAG Pipeline & AI
groq==0.9.0
httpx==0.27.0
pinecone-client==4.1.2
sentence-transformers==3.0.1
torch==2.3.1+cpu
langchain==0.2.6
python-dotenv==1.0.1
rank-bm25==0.2.2

# Utilities & Dependencies
numpy==1.26.4
urllib3==2.2.2
